backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Enterprise FastAPI backend — MySQL + RBAC + AuditLog"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi[standard]>=0.114.2,<1.0.0",
    "python-multipart>=0.0.7,<1.0.0",
    "email-validator>=2.1.0,<3.0.0",
    "tenacity>=8.2.3,<10.0.0",
    "pydantic>=2.0",
    "alembic>=1.12.1,<2.0.0",
    "aiomysql>=0.2.0",
    "sqlmodel>=0.0.21,<1.0.0",
    "pydantic-settings>=2.2.1,<3.0.0",
    "sentry-sdk[fastapi]>=2.0.0,<3.0.0",
    "pyjwt>=2.8.0,<3.0.0",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "jinja2>=3.1.4,<4.0.0",
    "emails>=0.6,<1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0,<2.0.0",
    "ruff>=0.2.2,<1.0.0",
    "coverage>=7.4.3,<9.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py311"
exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "W191", "B904"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
